+++ MapboxSearch/place-autocomplete/src/test/java/com/mapbox/search/autocomplete/PlaceAutocompleteImplTest.kt
@@ -163,7 +163,7 @@ internal class PlaceAutocompleteImplTest {
         }
 
         val response = runBlocking {
-            placeAutocomplete.suggestions(TEST_POINT)
+            placeAutocomplete.reverse(TEST_POINT)
         }
 
         assertEquals(TEST_AUTOCOMPLETE_SUGGESTIONS, response.value)
@@ -181,7 +181,7 @@ internal class PlaceAutocompleteImplTest {
         }
 
         val response = runBlocking {
-            placeAutocomplete.suggestions(TEST_POINT)
+            placeAutocomplete.reverse(TEST_POINT)
         }
 
         assertSame(error, response.error)
@@ -212,7 +212,7 @@ internal class PlaceAutocompleteImplTest {
             types = listOf(QueryType.POI, QueryType.ADDRESS, QueryType.STREET),
         )
 
-        runBlocking { placeAutocomplete.suggestions(TEST_POINT, options) }
+        runBlocking { placeAutocomplete.reverse(TEST_POINT, options) }
         coVerify { searchEngine.search(coreOptions) }
         verify(exactly = 1) { activityReporter.reportActivity(eq("place-autocomplete-reverse-geocoding")) }
     }
@@ -228,7 +228,7 @@ internal class PlaceAutocompleteImplTest {
             types = null
         )
 
-        runBlocking { placeAutocomplete.suggestions(TEST_POINT, options) }
+        runBlocking { placeAutocomplete.reverse(TEST_POINT, options) }
 
         assertEquals(ALL_TYPES, slotOptions.captured.types)
     }
@@ -244,7 +244,7 @@ internal class PlaceAutocompleteImplTest {
             types = emptyList()
         )
 
-        runBlocking { placeAutocomplete.suggestions(TEST_POINT, options) }
+        runBlocking { placeAutocomplete.reverse(TEST_POINT, options) }
 
         assertEquals(ALL_TYPES, slotOptions.captured.types)
     }
@@ -272,7 +272,7 @@ internal class PlaceAutocompleteImplTest {
         const val TEST_QUERY = "Test query"
 
         val TEST_BASE_SUGGESTIONS: List<BaseSearchSuggestion> = listOf(
-            createTestBaseSearchSuggestion(testBaseRawSearchSuggestionWithoutCoordinates)
+            createTestBaseSearchSuggestion(testBaseRawSearchSuggestionWithoutCoordinates.copy(center = Point.fromLngLat(10.0, 15.0)))
         )
 
         val TEST_BASE_RESULTS: List<BaseSearchResult> = listOf(testBaseResult)